--- mobileapp/utils/do-stuff.js
@@ -1,19 +1,0 @@
-/**
- * Module dependencies
- */
-
-// N/A
-
-
-
-/**
- * doStuff()
- *
- * Do something and return a message.
- *
- * @returns {String}
- */
-module.exports = function doStuff (){
-  // ...
-  return 'oh hey';
-};
